# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "查询用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "必填 页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "必填 页面大小",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "选填 用户类型",
                        "name": "admin_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "选填 用户级别",
                        "name": "level",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "选填 城市",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "选填 用户 id",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserV1ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus500"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "新增用户",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserV1CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserV1CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus500"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{user_id}": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户相关"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserV1UpdateBodyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserV1UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus400"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus401"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorStatus500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "resp.ErrorStatus400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 3
                },
                "message": {
                    "type": "string",
                    "example": "参数错误"
                }
            }
        },
        "resp.ErrorStatus401": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 16
                },
                "message": {
                    "type": "string",
                    "example": "未授权"
                }
            }
        },
        "resp.ErrorStatus404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 5
                },
                "message": {
                    "type": "string",
                    "example": "资源未找到"
                }
            }
        },
        "resp.ErrorStatus500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 10
                },
                "message": {
                    "type": "string",
                    "example": "服务器错误"
                }
            }
        },
        "user.UserStruct": {
            "type": "object",
            "properties": {
                "admin_type": {
                    "type": "integer"
                },
                "city": {
                    "type": "integer"
                },
                "credential_number": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sso_name": {
                    "type": "string"
                }
            }
        },
        "user.UserV1CreateRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "integer"
                },
                "credential_number": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sso_name": {
                    "type": "string"
                }
            }
        },
        "user.UserV1CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "user.UserV1ListResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "user_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserStruct"
                    }
                }
            }
        },
        "user.UserV1UpdateBodyRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "integer"
                },
                "credential_number": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "sso_name": {
                    "type": "string"
                }
            }
        },
        "user.UserV1UpdateResponse": {
            "type": "object"
        }
    }
}